# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CultureConnect API",
        "description": "API documentation for CultureConnect application - A cultural events and experiences platform",
        "contact": {
            "email": "support@cultureconnect.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard statistics",
                "description": "Get comprehensive statistics for the admin dashboard",
                "operationId": "d3c2f590b8522ac2b85b18cd681da4ff",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "overview": {
                                                    "properties": {
                                                        "total_users": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "active_users": {
                                                            "type": "integer",
                                                            "example": 120
                                                        },
                                                        "total_events": {
                                                            "type": "integer",
                                                            "example": 25
                                                        },
                                                        "total_orders": {
                                                            "type": "integer",
                                                            "example": 300
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "recent_activity": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized. Admin access required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}